Bug fixes:
Files now copied, not moved to processing location. This will prevent WinSCP from resynching them across to MASSIVE.
If any stdout or an error is detected during command execution this is captured and stored in the files "errors.txt" and  "output.txt". These files are created in the ExperimentID folder.